requirements.txt
altair
pandas
streamlit
requests
psycopg2-binary